Split SvgTransform
