+++ Source/Transforms/SvgTransform.Drawing.cs
@@ -1,0 +1,43 @@
+#if !NO_SDC
+using System.Drawing.Drawing2D;
+
+namespace Svg.Transforms
+{
+    public abstract partial class SvgTransform
+    {
+        public abstract Matrix Matrix { get; }
+
+        #region Equals implementation
+
+        public override bool Equals(object obj)
+        {
+            var other = obj as SvgTransform;
+            if (other == null)
+                return false;
+
+            return Matrix.Equals(other.Matrix);
+        }
+
+        public override int GetHashCode()
+        {
+            return Matrix.GetHashCode();
+        }
+
+        public static bool operator ==(SvgTransform lhs, SvgTransform rhs)
+        {
+            if (ReferenceEquals(lhs, rhs))
+                return true;
+            if (ReferenceEquals(lhs, null) || ReferenceEquals(rhs, null))
+                return false;
+            return lhs.Equals(rhs);
+        }
+
+        public static bool operator !=(SvgTransform lhs, SvgTransform rhs)
+        {
+            return !(lhs == rhs);
+        }
+
+        #endregion
+    }
+}
+#endif
